Created new Test class for the VHDL flavour of MetaMachine.
